requirements.txt
google-generativeai
python-dotenv
tenacity
